Fix dbus race condition during startup
Usually, razor-session starts up really fast, and by the time time
modules start using the D-Bus interface (for canLogout), it is already
registered. However, in the debug.log I've found that in some rare cases
modules fail to connect to the iterface because it's not yet registered,
which would result in a missing Logout item.