config: auto-detect succesfully kwin on recent kwin versions
Recently kde decided to split kwin's binary into two, kwin_x11 and
kwin_wayland in order to prepare the migration to wayland.

So to detect properly kwin on recent installations we need to look for
the binary kwin_x11 (and later kwin_wayland when LxQt gets ported to
wayland).

I've left the entry for kwin as it doesn't seem to interfere and allows
auto-detection of kwin to work on older intallations as well.