lxqt-leave: Handles fallback icons the proper way
We were setting the icons in the ui file. The generated code uses
QIcon::hasThemeIcon() to check the icon existence. But
QIcon::hasThemeIcon() doesn't take in account the fallbacks. The fallback
icons are never used and we end up with no icon.

Simply using QIcon::fromTheme() in the source code, fixes the issue.