Added `execDesktopFile` to lxqt-session's D-Bus adaptor (#579)
The reason is that we need a way of launching apps under the session tree with shortcuts provided by a Wayland compositor, which is itself outside that tree.

Previously, we recommended launching them with pcmanfm-qt, but apart from the annoying execution prompts, complications were possible. For example, if the desktop module wasn't running, a hidden pcmanfm-qt process would be started, and enabling the desktop module later would result in a desktop outside the session tree.

With this patch, apps can be launched by

```sh
qdbus org.lxqt.session /LXQtSession execDesktopFile DESKTOP_FILE_NAME_OR_PATH
```